INTYGFV-16761: Updated so that CS returns int-values instead of long-values..

+++ web/src/main/java/se/inera/intyg/webcert/web/csintegration/integration/dto/StatisticsForUnitDTO.java
@@ -30,8 +30,8 @@ import se.inera.intyg.webcert.web.csintegration.integration.dto.StatisticsForUni
 @Builder
 public class StatisticsForUnitDTO {
 
-    long draftCount;
-    long unhandledMessageCount;
+    int draftCount;
+    int unhandledMessageCount;
 
     @JsonPOJOBuilder(withPrefix = "")
     public static class StatisticsForUnitDTOBuilder {
